# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TempMail API - disposable email service
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Inbox {
  address: string;
  token: string;
  createdAt: string;
  expiresAt: string;
}

export interface EmailPreview {
  id: number;
  toAddress: string;
  fromAddress: string;
  subject: string;
  preview: string;
  hasAttachments: boolean;
  receivedAt: string;
}

export interface Email {
  id: number;
  toAddress: string;
  fromAddress: string;
  subject: string;
  textBody: string;
  htmlBody: string;
  hasAttachments: boolean;
  receivedAt: string;
}

export interface InboxWithEmails {
  address: string;
  createdAt: string;
  expiresAt: string;
  emails: EmailPreview[];
}

export type DomainStatus = (typeof DomainStatus)[keyof typeof DomainStatus];

export const DomainStatus = {
  active: "active",
  paused: "paused",
  blocked: "blocked",
} as const;

export interface Domain {
  id: number;
  name: string;
  status: DomainStatus;
  isPublic: boolean;
  emailCount: number;
  createdAt: string;
}

export interface DnsCheck {
  domain: string;
  mxValid: boolean;
  mxRecords: string[];
  checkedAt: string;
}

export type AdPlacement = (typeof AdPlacement)[keyof typeof AdPlacement];

export const AdPlacement = {
  header: "header",
  sidebar: "sidebar",
  inbox_top: "inbox_top",
  email_body: "email_body",
} as const;

export interface Ad {
  id: number;
  name: string;
  placement: AdPlacement;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  linkUrl?: string | null;
  isActive: boolean;
  impressions: number;
  createdAt: string;
}

export interface DashboardStats {
  totalEmails: number;
  emailsToday: number;
  activeInboxes: number;
  totalDomains: number;
  activeDomains: number;
  totalAds: number;
  activeAds: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  email_received: "email_received",
  inbox_created: "inbox_created",
  domain_added: "domain_added",
  ad_created: "ad_created",
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  message: string;
  timestamp: string;
}

export interface TimeBucket {
  hour: string;
  count: number;
}

export type CreateCustomInboxBody = {
  localPart: string;
  domainId: number;
};

export type ListAllEmailsParams = {
  limit?: number;
  domainId?: number;
  search?: string;
};

export type CreateDomainBody = {
  name: string;
  isPublic?: boolean;
};

export type UpdateDomainBodyStatus =
  (typeof UpdateDomainBodyStatus)[keyof typeof UpdateDomainBodyStatus];

export const UpdateDomainBodyStatus = {
  active: "active",
  paused: "paused",
  blocked: "blocked",
} as const;

export type UpdateDomainBody = {
  status?: UpdateDomainBodyStatus;
  isPublic?: boolean;
};

export type CreateAdBodyPlacement =
  (typeof CreateAdBodyPlacement)[keyof typeof CreateAdBodyPlacement];

export const CreateAdBodyPlacement = {
  header: "header",
  sidebar: "sidebar",
  inbox_top: "inbox_top",
  email_body: "email_body",
} as const;

export type CreateAdBody = {
  name: string;
  placement: CreateAdBodyPlacement;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  linkUrl?: string | null;
  isActive?: boolean;
};

export type UpdateAdBodyPlacement =
  (typeof UpdateAdBodyPlacement)[keyof typeof UpdateAdBodyPlacement];

export const UpdateAdBodyPlacement = {
  header: "header",
  sidebar: "sidebar",
  inbox_top: "inbox_top",
  email_body: "email_body",
} as const;

export type UpdateAdBody = {
  name?: string;
  placement?: UpdateAdBodyPlacement;
  content?: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  linkUrl?: string | null;
  isActive?: boolean;
};