# lib/api-zod/src/generated/types/domain.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TempMail API - disposable email service
 * OpenAPI spec version: 0.1.0
 */
import type { DomainStatus } from "./domainStatus";

export interface Domain {
  id: number;
  name: string;
  status: DomainStatus;
  isPublic: boolean;
  emailCount: number;
  createdAt: Date;
}